backend/requirements.txt
flask
flask-cors
flask-pymongo
python-dotenv
